feat: add dashscope and openai image client example
