+++ spring-ai-alibaba-image-example/dashscope-image/src/main/resources/application.yml
@@ -1,0 +1,10 @@
+server:
+  port: 10008
+
+spring:
+  application:
+    name: spring-ai-alibaba-dashscope-image-example
+
+  ai:
+    dashscope:
+      api-key: ${AI_DASHSCOPE_API_KEY}
