+++ spring-ai-alibaba-image-example/openai-image/README.md
@@ -1,0 +1,5 @@
+# Spring AI Alibaba OpenAI Image Example
+
+This is a Image client example for use DashScope LLMs by Spring AI Alibaba.
+
+TODO: add more info for ImageClient.
