+++ spring-ai-alibaba-mcp-example/filesystem/pom.xml
@@ -2,17 +2,20 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>3.3.6</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<relativePath/>
 	</parent>
+
 	<groupId>com.example</groupId>
 	<artifactId>ai-mcp-fileserver</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>Spring AI - Model Context Protocol - Fileserver</name>
 	<description>Simple AI Application using MCP client to chat with FileServer</description>
+
 	<properties>
 		<java.version>17</java.version>
 		<spring.ai.alibaba>1.0.0-M5.1-SNAPSHOT</spring.ai.alibaba>
