+++ spring-ai-alibaba-observability-example/observability-example/src/main/java/com/alibaba/cloud/ai/example/observability/FileSpanExporterAutoConfiguration.java
@@ -1,12 +1,11 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+ * Copyright 2025 the original author or authors.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.alibaba.cloud.ai.example.observability;
 
 import com.alibaba.cloud.ai.example.observability.exporter.oltp.OtlpFileSpanExporter;
@@ -28,6 +28,7 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ConditionalOnClass({OtlpFileSpanExporter.class})
 public class FileSpanExporterAutoConfiguration {
+
 	@Bean
 	@ConditionalOnMissingBean
 	@ConditionalOnEnabledTracing
@@ -35,4 +36,5 @@ public class FileSpanExporterAutoConfiguration {
 		OtlpFileSpanExporterProvider provider = new OtlpFileSpanExporterProvider();
 		return (OtlpFileSpanExporter)provider.createExporter(null);
 	}
+
 }
