--- spring-ai-alibaba-structured-example/output-parser-example/src/main/java/com/alibaba/cloud/ai/example/outparser/OutputParserExampleApplication.java
@@ -1,13 +1,0 @@
-package com.alibaba.cloud.ai.example.outparser;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OutputParserExampleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OutputParserExampleApplication.class, args);
-	}
-
-}
