--- spring-ai-alibaba-structured-example/output-parser-example/src/main/resources/application.yml
@@ -1,7 +1,0 @@
-spring:
-  application:
-    name: output-parser-example
-
-  ai:
-    dashscope:
-      api-key: ${AI_DASHSCOPE_API_KEY}
