+++ spring-ai-alibaba-structured-example/pom.xml
@@ -16,9 +16,8 @@
    limitations under the License.
 -->
 
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xmlns="http://maven.apache.org/POM/4.0.0"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
@@ -28,19 +27,39 @@
 		<relativePath>../pom.xml</relativePath>
 	</parent>
 
-	<version>${revision}</version>
 	<artifactId>spring-ai-alibaba-structured-example</artifactId>
-	<packaging>pom</packaging>
-
+	<version>0.0.1-SNAPSHOT</version>
+	<name>Spring AI Alibaba Structured Example</name>
 	<description>Spring AI Alibaba Structured Example</description>
-	<name>Spring AI Alibaba Structured Examples</name>
 
-	<modules>
-		<!--<module>dashscope</module>-->
-	</modules>
+	<properties>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
+		<maven-deploy-plugin.version>3.1.1</maven-deploy-plugin.version>
+
+		<!-- Spring AI -->
+		<spring-ai-alibaba.version>1.0.0-M3.3</spring-ai-alibaba.version>
+	</properties>
+
+	<dependencies>
+		<dependency>
+			<groupId>com.alibaba.cloud.ai</groupId>
+			<artifactId>spring-ai-alibaba-starter</artifactId>
+			<version>${spring-ai-alibaba.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-web</artifactId>
+		</dependency>
+	</dependencies>
 
 	<build>
 		<plugins>
+			<plugin>
+				<groupId>org.springframework.boot</groupId>
+				<artifactId>spring-boot-maven-plugin</artifactId>
+			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-deploy-plugin</artifactId>
@@ -51,5 +70,16 @@
 			</plugin>
 		</plugins>
 	</build>
+
+	<repositories>
+		<repository>
+			<id>spring-milestones</id>
+			<name>Spring Milestones</name>
+			<url>https://repo.spring.io/milestone</url>
+			<snapshots>
+				<enabled>false</enabled>
+			</snapshots>
+		</repository>
+	</repositories>
 
 </project>
